use mockApp for tests
